Day11:项目实战
1.完成学员方面的所有接口
2.新增班级和部门删除时判断有没有人

+++ tlias-web-management/src/main/java/com/itheima/controller/ClazzController.java
@@ -1,6 +1,10 @@
 package com.itheima.controller;
 
-import com.itheima.pojo.*;
+import com.itheima.exception.ClazzNotEmptyException;
+import com.itheima.pojo.Clazz;
+import com.itheima.pojo.ClazzQueryParam;
+import com.itheima.pojo.PageResult;
+import com.itheima.pojo.Result;
 import com.itheima.service.ClazzService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,6 +20,7 @@ public class ClazzController {
     @Autowired
     private ClazzService clazzService;
 
+    //查询所有班级
     @GetMapping("/list")
     public Result list() {
         log.info("查询全部班级数据");
@@ -23,6 +28,7 @@ public class ClazzController {
         return Result.success(clazzList);
     }
 
+    //添加班级
     @PostMapping
     public Result add(@RequestBody Clazz clazz) {
         log.info("新增班级：{}", clazz);
@@ -30,6 +36,7 @@ public class ClazzController {
         return Result.success();
     }
 
+    //班级列表查询
     @GetMapping
     public Result page(ClazzQueryParam clazzQueryParam) {
         log.info("Clazz分页查询：{}", clazzQueryParam);
@@ -37,13 +44,21 @@ public class ClazzController {
         return Result.success(pageResult);
     }
 
+    //删除班级
     @DeleteMapping("/{id}")
     public Result delete(@PathVariable Integer id) {
+
+        int studentCount = clazzService.getStudentCountByClazzId(id);
+        if (studentCount > 0) {
+            throw new ClazzNotEmptyException("对不起, 该班级下有学生, 不能直接删除~");
+        }
+
         log.info("删除的id班级为：{}", id);
         clazzService.deleteById(id);
         return Result.success();
     }
 
+    //修改班级
     @PutMapping
     public Result update(@RequestBody Clazz clazz) {
         log.info("修改班级：{}", clazz);
@@ -51,6 +66,7 @@ public class ClazzController {
         return Result.success();
     }
 
+    //根据ID查询
     @GetMapping("/{id}")
     public Result getInfo(@PathVariable Integer id) {
         log.info("ID查询班级：{}", id);
